[refactor] Move dependencies in back-end to subfolder
