+++ back-end/src/internal/dependencies/k8s_client.py
@@ -2,7 +2,7 @@
 from kubernetes.client import ApiClient, Configuration
 from kubernetes.config import ConfigException, load_incluster_config, load_kube_config
 
-from ..config.config import config
+from ...config.config import config
 
 
 def get_k8s_client() -> ApiClient:
